backend/requirements.txt
fastapi==0.115.6
uvicorn==0.34.0
httpx
python-dotenv
python-multipart
openai
pydantic
sqlalchemy
google-auth
python-jose[cryptography]
argon2-cffi
passlib
email-validator
psycopg2-binary
